==> data_science_and_statistic_with_python/00-basic_python.ipynb <==
# **Introdução**

O objetivo desse projeto é de apresentar os comandos básicos de Python para aplicar em análise de dados.

print ("Olá Mundo")

mensagem = 'Olá Mundo'

print (mensagem)

mensagem

nome = input ('Qual é o seu nome?')

print ('Meu nome é:',  nome)

nome = input ('Qual é o seu nome?')
print ('Meu nome é:',  nome)


# **Operadores Matemáticos**

n1 = 7

n2 = 2

s = n1 + n2
print(s)

s = s + n2
print (s)

s += n2
print(s)

sub = n1 - n2
sub

mult = n1 * n2
mult

div = n1 / n2
div

pot = n1**n2
pot

pot2 = pow (n1,n2)
pot2

divi = n1 // n2
divi

rest = n1 % n2
rest

# **Importações de bibliotecas, pacotes e módulos**

raiz_quadrada = sqrt (25)

import math


math.

n = 18

raiz = math.sqrt(n)
print ('A raiz de {} é igual a {:.2f}'.format (n, raiz))

print(raiz)

raiz2 = n ** 0.5
print (raiz2)

print ('{:.3f}'.format(raiz2))

from math import sqrt
raiz3=sqrt(n)
print ('{:.2f}'.format(raiz3))

from math import sqrt, factorial

fat = factorial(5)       # fatorial: 4! = 4.3.2 = 24
                         #           5! = 5.4.3.2 = 120
fat

fatorial_raiz = factorial(sqrt(16))
fatorial_raiz

import random
num = random.randint (1,10)
print (num)

random.

from random import random, randint

random?

randint?

num2 = random()
num2

# **Estrutura condicional**

if        
else        
elif

nota = 5

if nota >= 6:
    print('Aprovado')
else:
    print('Reprovado')

if nota >= 6:
    print('Aprovado')
elif nota >= 5 and nota< 6:
    print('Recuperação')
else:
    print('Reprovado')

# **Estrutura de Repetição**

for    
while    


for c in range(0,4):
    print(c)

lista = [1, 2, 3, 4, 10]
for numero in lista:
  mult = numero * 2
  print(mult)

soma = 0
cont = 0
for c in range (1, 6):
    num = int(input('digite o {}o valor: '.format(c)))
    if num % 2 == 0:
        soma = soma + num 
        cont = cont + 1
print('Você informou {} números pares e a soma deles é igual a {}'.format(cont, soma))

c = 1
while c <= 10:
    s = c + 10
    print(s)
    c = c + 1
print('FIM')

# **Listas, Tuplas e Dicionários**

Listas [ ]      
Tuplas ( )     
Dicionários { }     

**LISTAS**

num = [2, 5, 7, 9]

num[3] = 4   # no lugar do elemento na posição 3 (número 9) vai ser colocado o número 4

num

num.append(8) #inclui o número 8
num

num.sort() #coloca em ordem crescente
num

num.insert(2, 0) # adiciona o zero na posição 2
num

num.remove(7) #remove o elemento 7
num

num.pop(3) #remove o elemento na posição 3
num

**TUPLAS**

Tupla é imutável.

num2 = (2, 5, 7, 9)

num2[3] = 4 

num2.append(8) #inclui o número 8
num2

**DICIONÁRIOS**

pessoas = {'nome': 'Luciano', 'sexo': 'masculino', 'idade': 46}  # número não coloca aspas
print (pessoas)

pessoas['time'] = 'Corinthians'
print(pessoas)

del pessoas['idade']
print(pessoas)

# **Funções**

def identificacao(nome,idade):
   print('Olá',nome,'\nVocê é jovem, tem apenas',idade, 'anos')

identificacao('Luciano', 46)

def maior (x,y):
  if x < y:
    print ('O maior número é', y)
  elif x == y:
    print ('Os números são iguais')
  else:
    print ('O maior número é', x)

maior (11,16)

def pitagoras (cat1, cat2, hip):
  if hip == '?':
    hip = (cat1**2+cat2**2)**(1/2)
    print ('A hipotenusa é', hip)
  elif cat1 == '?':
    cat1 = (hip**2-cat2**2)**(1/2)
    print ('O cateto é', cat1)
  elif cat2 == '?':
    cat2 = (hip**2-cat1**2)**(1/2)
    print ('O cateto é', cat2)

pitagoras (13,5,'?')

## **Função lambda e map**

Forma simplificada de escrever funções

def area_quadrado (L):
  area = L**2
  print(area)

area_quadrado(4)

area_quadrado2 = lambda x: x**2

area_quadrado2(7)

area_retangulo = lambda b, h: b*h

area_retangulo(4,7)

# Uma vantagem é quando se utiliza juntamente com outros métodos
L = [4,5,6,7,11,9,10]
areas = list(map(lambda x: x**2, L))
print(areas)

# **List Comprehensions**

Otimiza a utilização de listas e diminui linhas de código

lista=[]
for valor in range(5):
  lista.append(valor+10)
print(lista)


lista = [valor+10 for valor in range(5)]
print(lista)


lista = []
for numero in range (1,30):
  if numero % 4 == 0:
    lista.append(numero)
print(lista)


multiplos4 = [numero for numero in range(1,30) if numero % 4 == 0]
print(multiplos4)


conceito = ['azul' if nota >= 6 else 'vermelha' for nota in range(1,11)]
print(conceito)

# **Vetores (arrays) e matrizes**

Matriz   
1    3    5    7   
2    4    6    8   
0    1    2    3   

Vetor   
5  10  15  20   
ou   
3   
4   
5   
6   

vetor = [2,4,5,6]

type(vetor)

matriz = [[1,2,3],
          [4,5,6],
          [7,8,9]]

type(matriz)

print(matriz)

import numpy as np

vetor2 = np.array([1, -3, 4.2, 5])

type(vetor2)

print (vetor2)

matriz2 = np.ones((2,3))
matriz2

type(matriz2)

matriz3 = np.diag((2,4,6,8,10))
matriz3

type(matriz3)
