==> data_science_and_statistic_with_python/04-pandas_csv.ipynb <==
import pandas

data4 = pandas.read_csv(
    "classificados.csv",
    sep=",",
    encoding="iso-8859-1",
)
data4.shape

data4.head()

# Query de busca de notas de redação q conta os valores com notas 0
# nota da redação soma dos comp1 até 5
data4.query('NOTA_REDACAO == 0')["NOTA_REDACAO"].value_counts()

data4.query('NOTA_CH == 0')["NOTA_CH"].value_counts()


# Ciencias humanas
data4.query('NOTA_CH == 0')["NOTA_CH"].value_counts()

# Ciencias da natureza
data4.query('NOTA_CN == 0')["NOTA_CN"].value_counts()

# Linguagens e códigos
data4.query('NOTA_LC == 0')["NOTA_LC"].value_counts()

# MATEMÁTICA
data4.query('NOTA_MT == 0')["NOTA_MT"].value_counts()

# remover colunas usadas para filtro
test_classified2 = data4.drop(columns=["IN_TREINEIRO", "presenca_total"])

test_classified2.shape

test_classified2.head(2)

# salvar resultado com apenas os classificados
test_classified2.to_csv("csv/04-classificados_tratado.csv", encoding="iso-8859-1", index=False)
