==> data_science_and_statistic_with_python/18-19-pandas_csv.ipynb <==
import pandas

# encoding: codificação de caracteres, normalmente utiliza-se o iso-8859-1, utf-8, latin-1)
data = pandas.read_csv(
    "/media/wagner/HD-250GB2/Estatistica_Analise_Dados/csv/microdados_enem_2019_sp.csv",
    sep=";",
    encoding="iso-8859-1",
)

# [813772 rows x 44 columns]
print(data)

# visualizar o cabeçalho os 5 primeiros registros
# [5 rows x 44 columns]
data.head()

# especifica a quant. de registros
data.head(10)

# total da matriz L x C
data.shape

# Tipo dos dados por coluna
data.dtypes


data.shape

# excluir colunas "," no final quando forem mais de 1
data2 = data.drop(columns=["CO_MUNICIPIO_RESIDENCIA", 'CO_UF_RESIDENCIA','SG_UF_RESIDENCIA','CO_MUNICIPIO_NASCIMENTO', 'NO_MUNICIPIO_NASCIMENTO','CO_UF_NASCIMENTO','SG_UF_NASCIMENTO',
                              'TP_ANO_CONCLUIU','TP_ENSINO','CO_MUNICIPIO_ESC','CO_UF_ESC','SG_UF_ESC', 'TP_DEPENDENCIA_ADM_ESC','TP_LOCALIZACAO_ESC','TP_SIT_FUNC_ESC'],)
data2.shape

# problema colunas foram importadas com uma casa decimal a mais
# loc localizar ":" todas as linhas da coluna xxx e dividiro por 10
data2.loc[:, "NU_NOTA_CH"] /= 10
data2.loc[:, "NU_NOTA_CN"] /= 10
data2.loc[:, "NU_NOTA_LC"] /= 10
data2.loc[:, "NU_NOTA_MT"] /= 10
data2.head(3)

# Renomear coluna
data2 = data2.rename(columns={"NU_NOTA_REDACAO":"NOTA_REDACAO"})
data2.head(3)

# renomear várias colunas
data2 = data2.rename(columns={'NU_NOTA_CN': 'NOTA_CN','NU_NOTA_CH': 'NOTA_CH','NU_NOTA_LC': 'NOTA_LC','NU_NOTA_MT': 'NOTA_MT', 'NU_NOTA_COMP1': 'COMP1','NU_NOTA_COMP2': 'COMP2',
                            'NU_NOTA_COMP3': 'COMP3','NU_NOTA_COMP4': 'COMP4', 'NU_NOTA_COMP5':'COMP5', 'NU_IDADE': 'IDADE','TP_SEXO': 'SEXO', 'TP_COR_RACA': 'RACA','Q025': 'INTERNET',
                            'TP_ESCOLA':'ESCOLA'})
data2.head(3)

# renomear itens de uma coluna
# troca na coluna RACA o código num. pela nomenclatura
data2["RACA"] = data2["RACA"].replace({0:"nao_declarado", 1:"branca", 2:"preta", 3:"parda", 4:"amarela", 5:"indigena"}) 
# OU
# dados1['RACA'] = dados1['RACA'].replace([0,1,2,3,4,5], ['Não Declarado','Branca','Preta','Parda','Amarela','Indígena'])

data2["TP_LINGUA"] = data2["TP_LINGUA"].replace({0:"Inglês", 1:"Espanhol"})
data2["ESCOLA"] = data2["ESCOLA"].replace({1:"não_respondeu", 2:"pública", 3:"privada", 4:"exterior"})
data2["INTERNET"] = data2["INTERNET"].replace({"A":"Não", "B":"Sim"})

data2.head(3)

#  contador (contagem dos valores idade e quant. de vezes q aparece)
data2["IDADE"].value_counts()

# cotador em ordem crescente
data2["IDADE"].value_counts().sort_index()

# QUERY contagem dos menores de 12 anos por cidade
age_small_12 = data2.query("IDADE < 12")["NO_MUNICIPIO_RESIDENCIA"].value_counts()
print(age_small_12)

# Localizar maiores de 11 dados.coluna (pega os dados completos com todas colunas)
age_more_11 = data2.loc[data2.IDADE > 11]
age_more_11.head(3)

# Localizar Contar e Ordenar maiores de 11 dados.coluna
age_more_11_order = age_more_11["IDADE"].value_counts().sort_index()
print(age_more_11_order)

# Treineiro (1 quem faz a prova apenas para treino 93mil)
student_test = age_more_11.query("IN_TREINEIRO == 1")["IN_TREINEIRO"].value_counts()
print(student_test)

# Localizar variável.coluna
student_test2 = age_more_11.loc[age_more_11.IN_TREINEIRO == 1]
student_test2.head(3)

student_test2.shape

# Criar arquivo CSV com resultados
# index False ou True para aparecer a coluna do index do pandas
student_test2.to_csv("csv/02-treineiros_enem_2019_sp.csv", encoding="iso-8859-1", index=False)
