==> data_science_and_statistic_with_python/22-pandas_csv.ipynb <==
import pandas

data3 = pandas.read_csv(
    "treineiros_enem_2019_sp.csv",
    sep=",",
    encoding="iso-8859-1",
)
data3.shape

data3.head(3)

#  analisar presença 1 ok, 0 faltaram, 2 eliminados
# CN (Ciencia natureza) / prova dia 01
data3["TP_PRESENCA_CN"].value_counts()

# ch ciencias humanas / prova dia 02
data3["TP_PRESENCA_CH"].value_counts()

# lc linguagem e código / prova dia 02
data3["TP_PRESENCA_LC"].value_counts()

# MT matemática / prova dia 01
data3["TP_PRESENCA_MT"].value_counts()

# status redação 1 nenhum problema
data3["TP_STATUS_REDACAO"].value_counts().sort_index()

# verificar se existem valores nulos "em branco", não inclui o zero.
# soma a quantidade de registros
data3["NOTA_CN"].isnull().sum()

# Agrupar em uma variável todas as colunas com notas de provas
# soma do total de faltas registros individual de cada coluna
# as q têm o mesmo número foram aplicadas no mesmo dia
# tests passado como variável no array do data3 para soma
tests = ["NOTA_REDACAO", "NOTA_CN", "NOTA_MT", "NOTA_LC", "NOTA_CN"]
data3[tests].isnull().sum()

# soma a quantidade de presenças, mesma quant. mesmo dia.
total_attendances = ["TP_PRESENCA_CN", "TP_PRESENCA_CH", "TP_PRESENCA_LC", "TP_PRESENCA_MT"]
data3[total_attendances].sum()

# Cria uma Coluna com a soma dos valores de presenças se for igual a 4 participou dos 2 dias
# soma a quant. se o valor for igual a 1
total_attendances2 = ["TP_PRESENCA_CN", "TP_PRESENCA_CH", "TP_PRESENCA_LC", "TP_PRESENCA_MT"]
data3["presenca_total"] = data3[total_attendances2].sum(axis=1)
data3.head(5)

# contagem dos totais de presenças por valores recebidos 6-desclassificados
data3["presenca_total"].value_counts().sort_index()

# buscar apenas os q têm presença total igual a 4
total_attendances3 = data3.loc[data3.presenca_total == 4]

# problema ainda existem q estavam presentes mas forma desclassificados
# "TP_PRESENCA_CN", "TP_PRESENCA_CH", "TP_PRESENCA_LC", "TP_PRESENCA_MT"
total_attendances3["TP_PRESENCA_CN"].value_counts().sort_index()

total_attendances3["TP_PRESENCA_LC"].value_counts().sort_index()

test_classified = total_attendances3.loc[total_attendances3.TP_PRESENCA_CN == 1]

# "TP_PRESENCA_CN", "TP_PRESENCA_CH", "TP_PRESENCA_LC", "TP_PRESENCA_MT"
test_classified["TP_PRESENCA_CN"].value_counts().sort_index()


test_classified.shape

# verificar se ainda existem provas com valores ausentes conforme feito antes:
tests = ["NOTA_REDACAO", "NOTA_CN", "NOTA_MT", "NOTA_LC", "NOTA_CN"]
test_classified[tests].isnull().sum()

# salvar resultado com apenas os classificados
test_classified.to_csv("csv/03-classificados.csv", encoding="iso-8859-1", index=False)
