==> data_science_and_statistic_with_python/25-amostragem_simples.ipynb <==
import numpy
import pandas

# POPULAÇÃO:
data = pandas.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")

data.head(2)

data.shape

# AMOSTRAGEM: Cidade de Campinas
campinas = data.loc[data.NO_MUNICIPIO_RESIDENCIA == "Campinas"]
campinas.shape

# AMOSTRAGEM SIMPLES: (Seleção aleatória sem filtro)
# Criar uma "semente" de valor aleatório que será sempre igual
# gera sempre a mesma resposta (ver index panda sempre o mesmo),
# senão for definido cada vez q rodar gera uma amostragem diferente
numpy.random.seed(1)



# sample (aleatório)
# frac (fração capturar uma amostragem aleatória de 30% da população)
amostragem_simples_percentual = campinas.sample(frac=0.30)
amostragem_simples_percentual.shape

# frac (fração capturar uma amostragem aleatória numérica de xxx da população)
amostragem_simples_numerica = campinas.sample(595)
amostragem_simples_numerica.shape

amostragem_simples_numerica.head(3)

# Alterar o número mantém as amostragens iguais em cima e muda as próximas variáveis
numpy.random.seed(2)


# OU
# AMOSTRAGEM SIMPLES (outra forma mais complexa)
# Cria um array com números 0 (80%) e 1(20%) com a quant. de linhas da matriz principal
array_de_opcoes = [0,1]
array_binario_aleatorio = numpy.random.choice(array_de_opcoes, size=1982, replace=True, p=[0.8, 0.2])
# OU
# array_binario_aleatorio = numpy.random.choice(array_de_opcoes=[0,1], size=13198, replace=True, p=[0.8, 0.2])
array_binario_aleatorio


# 20% de 1
len(array_binario_aleatorio[array_binario_aleatorio == 1])

# Coleta a amostragem conforme 
amostra_campinas_binario_1 = campinas.loc[array_binario_aleatorio == 1]
amostra_campinas_binario_1.shape

amostra_campinas_binario_1.head(2)

amostragem_simples_percentual.to_csv("csv/05-amostragem_simples_percentual.csv", encoding="iso-8859-1", index=False)
amostragem_simples_numerica.to_csv("csv/05-amostragem_simples_numerica.csv", encoding="iso-8859-1", index=False)
