==> data_science_and_statistic_with_python/26-amostragem_sistematica.ipynb <==
import numpy
import pandas

# POPULAÇÃO:
data = pandas.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
data.shape

# AMOSTRAGEM: Cidade de Campinas
campinas = data.loc[data.NO_MUNICIPIO_RESIDENCIA == "Campinas"]
campinas.shape

# pegar um número aleatório de início de 0 até 10, 1 - quant. de números que vai pegar.
# se mudar o num. do seed muda o array de índece
numpy.random.seed(1)
begin = numpy.random.choice(10, 1)
begin

# AMOSTRAGEM SISTEMATICA: (Ordena EL numericamente / seleciona aleatoriamente usando uma seq. numerica)
# criar um array de índice dos EL que serão capturados
# (Valor inicial, valor max, e valor de salto de x em x )
array_numeros_amostras = numpy.arange(begin, 1982, 100)
array_numeros_amostras

# iloc (i de índice) / : (se refere a a todas as colunas)
amostra_sistematica = campinas.iloc[array_numeros_amostras,:]
amostra_sistematica.shape

amostra_sistematica.head(2)

# C/ Salto aleatório:
numpy.random.seed(2)
# begin (pegar um número aleatório de início de 0 até 10, e quant. de números que vai pegar "1")
# jump (Valor salto aleatório, saltar entre 100 e 150, e quant. de números que vai pegar "1" )

begin = numpy.random.choice(10, 1)
jump = numpy.random.randint(100, 150, 1)
print(begin, jump)

array_numeros_amostras = numpy.arange(begin, 1982, jump)
array_numeros_amostras

amostra_sistematica_2 = campinas.iloc[array_numeros_amostras, :]
amostra_sistematica_2.shape

amostra_sistematica_2.head(2)

amostra_sistematica.to_csv("csv/06-amostra_sistematica.csv", encoding="iso-8859-1", index=False)
amostra_sistematica_2.to_csv("csv/06-amostra_sistematica_2.csv", encoding="iso-8859-1", index=False)
