==> data_science_and_statistic_with_python/27-amostragem_estratificada.ipynb <==
import pandas
from sklearn.model_selection import train_test_split

# POPULAÇÃO:
data = pandas.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
data.shape

# AMOSTRAGEM: Cidade de Campinas
campinas = data.loc[data.NO_MUNICIPIO_RESIDENCIA == "Campinas"]
campinas.shape

# AMOSTRAGEM ESTRATIFICADA: (Divisão da pop. em grupos, e retira uma amostra desses grupos)
campinas["RACA"].value_counts()

# criar tabelas com amostragem de 10% de cada raça
branca = campinas.loc[campinas.RACA == "branca"].sample(frac=0.10)
parda = campinas.loc[campinas.RACA == "parda"].sample(frac=0.10)
preta = campinas.loc[campinas.RACA == "preta"].sample(frac=0.10)
amarela = campinas.loc[campinas.RACA == "amarela"].sample(frac=0.10)
nao_declarado = campinas.loc[campinas.RACA == "nao_declarado"].sample(frac=0.10)
indigena = campinas.loc[campinas.RACA == "indigena"].sample(frac=0.10)
branca.shape, parda.shape, preta.shape, amarela.shape, nao_declarado.shape, indigena.shape

# concatenar, juntar tabelas:
amostra_estratificada = pandas.concat([branca, parda, preta, amarela, nao_declarado, indigena])
amostra_estratificada.shape

amostra_estratificada.head(3)

amostra_estratificada.to_csv("csv/07-amostra_estratificada.csv", encoding="iso-8859-1", index=False)


# OU
# train_test_split
# axis 1 representa coluna, axis 0 linha
# vai extrair 10% dessa coluna
# ERRO se não tiver mais q 1 elemento da mesma raça
X_treino, X_teste, y_treino, amostra_estratificada_2 = train_test_split(campinas.drop('RACA',axis=1),campinas['RACA'], stratify=campinas['RACA'], test_size=0.10)
amostra_estratificada_2.shape, amostra_estratificada_2.head()


amostra_estratificada_2.value_counts()

# visulaizar os índices
amostra_estratificada_2.index

# caputra os indices
index_amostras = amostra_estratificada_2.index


# através dos índeces pega todos os campos da planilha original
amostra_estratificada_2_campos = campinas.loc[index_amostras, :]
amostra_estratificada_2_campos.shape

amostra_estratificada_2_campos.head(2)

# Gera uma planilha igual a anterior (199L x 28C)
amostra_estratificada_2_campos.to_csv("csv/07-amostra_estratificada_2_campos.csv", encoding="iso-8859-1", index=False)
