==> data_science_and_statistic_with_python/28-amostragem_conglomerado.ipynb <==
import numpy
import pandas

# POPULAÇÃO:
data = pandas.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
data.shape

# AMOSTRAGEM: Cidade de Campinas
campinas = data.loc[data.NO_MUNICIPIO_RESIDENCIA == "Campinas"]
campinas.shape

# AMOSTRAGEM CONGLOERADO "Agrupamentos" 
# (Divisão da pop. em grupos com características semelhantes, 
# Seleção aleatória de alguns grupos, analisa TODOS EL dos grupos selecionados)
# ex. pessoas brancas da mesma idade
brancas = campinas.loc[campinas.RACA == "branca"]
brancas.shape

amostras_todas_idades = brancas["IDADE"].value_counts()
amostras_todas_idades

# gera um índice para cada idade total de 8
amostras_todas_idades.index

# Gera um array binário do tamanho dos índices de idade com percentual de 50% cada
indice_array_binario = numpy.random.choice(a=[0,1],size=8, replace=True, p=[0.80, 0.20])
indice_array_binario

# Não foi definida semente cada vez vai pegar um índice diferente
amostragem_conglomerado = amostras_todas_idades.loc[indice_array_binario == 1]
amostragem_conglomerado

amostragem_conglomerado.index

# isin busca os códigos 
amos_conglomerado = brancas.loc[brancas['IDADE'].isin([17, 18])]
amos_conglomerado

amos_conglomerado.to_csv("csv/08-amostragem_conglomerado.csv", encoding="iso-8859-1", index=False)
