==> data_science_and_statistic_with_python/30-distribuicao_de_frequencia.ipynb <==
import pandas
from collections import Counter

# POPULAÇÃO:
data = pandas.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
data.shape

# AMOSTRAGEM: Cidade de Campinas
campinas = data.loc[data.NO_MUNICIPIO_RESIDENCIA == "Campinas"]
campinas.shape

# cria uma variável com a coluna idades da tab.
idades = campinas["IDADE"]
idades.shape, idades

# FREQUÊNCIA ABSOLUTA: (Totais por grupo de idade)
frequencia_absoluta = Counter(idades)
frequencia_absoluta

# Converte em dicionário e coloca em ordem crescente (False - Decrescente)
frequencia_absoluta = pandas.DataFrame.from_dict(frequencia_absoluta, orient="index").sort_index(ascending=True)
frequencia_absoluta

frequencia_absoluta.to_csv("csv/09-frequencia_absoluta.csv", encoding="iso-8859-1", index=False)

# FREQUENCIA RELATIVA (divisão da absoluta pelo total)
frequencia_relativa = frequencia_absoluta / frequencia_absoluta.sum()
frequencia_relativa

frequencia_relativa.to_csv("csv/09-frequencia_relativa.csv", encoding="iso-8859-1", index=False)

# FREQUENCIA RELATIVA PERCENTUAL (percentual de cada idade)
frequencia_relativa_percentual = frequencia_relativa * 100
frequencia_relativa_percentual

frequencia_relativa_percentual.to_csv("csv/09-frequencia_relativa_percentual.csv", encoding="iso-8859-1", index=False)
