==> data_science_and_statistic_with_python/36-medidas_dispersao_posicao.ipynb <==
import pandas as pd

# POPULAÇÃO:
data = pd.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
data.shape

# AMOSTRAGEM: Cidade de Campinas
campinas = data.loc[data.NO_MUNICIPIO_RESIDENCIA == "Campinas"]
campinas.shape

# MEDIDAS DE DISPERSÃO
# VARIÂNCIA:
data["IDADE"].var()

# DESVIO PADRÃO:
data["IDADE"].std()

# DESVIO ABSOLUTO:
data["IDADE"].mad()

data["IDADE"].describe()

# Notas de matemática estão mais dispersas que as notas de redação.
round(data.NOTA_MT.std(), 2) ,round(data.NOTA_REDACAO.std(), 2)

# Comparar apenas as colunas de notas
colunas_notas = ["NOTA_REDACAO", "NOTA_CN", "NOTA_MT", "NOTA_LC", "NOTA_CH"]
data[colunas_notas].describe()

# MEDIDAS DE POSIÇÃO
# Mínimo:
data.IDADE.min()

# 1º Quartil - 25%
data.IDADE.quantile(q=0.25)

# 2º Quartil - 50% (Mediana)
data.IDADE.median(), data.IDADE.quantile(q=0.5)

# 3º Quartil - 75%
data.IDADE.quantile(q=0.75)

# Máximo:
data.IDADE.max()

data["IDADE"].describe()
