==> data_science_and_statistic_with_python/39-probabilidade_eventos.ipynb <==
import pandas as pd

# POPULAÇÃO:
dataset_sp = pd.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
dataset_sp.shape

dataset_bauru = dataset_sp.loc[dataset_sp.NO_MUNICIPIO_RESIDENCIA == "Bauru"]
dataset_bauru.shape

dataset_guarulhos = dataset_sp.loc[dataset_sp.NO_MUNICIPIO_RESIDENCIA == "Guarulhos"]
dataset_guarulhos.shape

# FN Probabilidade de Ocorrer um Evento
# events_possible = número de Eventos possíveis dentro do espaço amostral
#  sample_space =  Espaço Amostral de todas possibilidades
# :.2f número de casas após a vírgula
def probability_event(events_possible, sample_space):
    result = (events_possible / sample_space)*100
    print("{:.2f}%".format(result))
    return result

# prob. de sortear um aluno de Bauru
probability_event(len(dataset_bauru), len(dataset_sp))

# prob. de sortear um aluno de Guarulhos
probability_event(len(dataset_guarulhos), len(dataset_sp))

# Dado de tirar o número 2:
probability_event(1, 6)

# Dado de tirar o número par:
probability_event(3, 6)

# FN Probabilidade de NÃO Ocorrer um Evento
# events_possible = número de Eventos possíveis dentro do espaço amostral
#  sample_space =  Espaço Amostral de todas possibilidades
# :.2f número de casas após a vírgula
def probability_of_not_event(events_possible, sample_space):
    result =  (1-(events_possible / sample_space))*100
    print("{:.2f}%".format(result))
    return result

# prob. de NÃO sortear um aluno de Bauru
probability_of_not_event(len(dataset_bauru), len(dataset_sp))

# prob. de NÃO sortear um aluno de Guarulhos
probability_of_not_event(len(dataset_guarulhos), len(dataset_sp))

# Dado de NÃO tirar o número 2:
probability_of_not_event(1, 6)

# Dado de NÃO tirar o número par:
probability_of_not_event(3, 6)

# FN Probabilidade de Ocorrer DOIS Eventos E
# events_possible = número de Eventos possíveis dentro do espaço amostral
#  sample_space =  Espaço Amostral de todas possibilidades
def probability_event(events_possible_a, sample_space_a, events_possible_b, sample_space_b):
    sample_space = sample_space_a * sample_space_b
    events_possible = events_possible_a + events_possible_b
    result = ((events_possible / sample_space)*)*100
    print("{:.2f}%".format(result))
    return result

# prob. de sortear um aluno de Bauru
probability_event(len(dataset_bauru), len(dataset_sp))
