==> data_science_and_statistic_with_python/40-probabilidade_eventos_E_OU_SE.ipynb <==
import pandas as pd

# POPULAÇÃO:
dataset_sp = pd.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
dataset_sp.shape

dataset_bauru = dataset_sp.loc[dataset_sp.NO_MUNICIPIO_RESIDENCIA == "Bauru"]
dataset_bauru.shape

dataset_guarulhos = dataset_sp.loc[dataset_sp.NO_MUNICIPIO_RESIDENCIA == "Guarulhos"]
dataset_guarulhos.shape

# FN Probabilidade União de Eventos AUB - OU (Mutuamente Exclusivos)
# Soma prob. dos eventos
# events_possible = número de Eventos possíveis dentro do espaço amostral
#  sample_space =  Espaço Amostral de todas possibilidades
def probability_union_event(events_possible_a, events_possible_b, sample_space):
    events_possible_a = events_possible_a / sample_space
    events_possible_b = events_possible_b / sample_space
    result =  (events_possible_a + events_possible_b)*100
    print("{:.2f}%".format(result))
    return result

# prob. de sortear um aluno de Bauru ou Guarulhos
probability_union_event(len(dataset_bauru), len(dataset_guarulhos), len(dataset_sp))

# Dados tirar os números x OU y
probability_union_event(1, 1, 6)

# FN Probabilidade União de Eventos AUB - OU (Mutuamente Exclusivos)
# Considerando a Hipótese que OU X pessoas participaram de ambos eventos
# Nessa caso soma prob. dos eventos e diminui a prob. dos mesmos juntos
# events_possible = número de Eventos possíveis dentro do espaço amostral
#  sample_space =  Espaço Amostral de todas possibilidades
def probability_union_event2(events_possible_a, events_possible_b, sample_space, both_events):
    events_possible_a = events_possible_a / sample_space
    events_possible_b = events_possible_b / sample_space
    both_events_possible = both_events / sample_space

    result =  (events_possible_a + events_possible_b - both_events_possible)*100
    print("{:.2f}%".format(result))
    return result

# prob. de sortear um aluno q fez vestibular em Bauru OU Guarulhos, OU nos dois.
both_events = 300
probability_union_event2(len(dataset_bauru), len(dataset_guarulhos), len(dataset_sp), both_events)


# FN Probabilidade Intersecção de Eventos A∩B - E (Dois Eventos Independentes)
# a prob. é igual o produto dos eventos
# events_possible = número de Eventos possíveis dentro do espaço amostral
#  sample_space =  Espaço Amostral de todas possibilidades
def probability_intersection_event(events_possible_a, events_possible_b, sample_space):
    events_possible_a = events_possible_a / sample_space
    events_possible_b = events_possible_b / sample_space

    result =  (events_possible_a * events_possible_b)*100
    print("{:.2f}%".format(result))
    return result

# prob. de sortear em sequência um aluno q fez vestibular em Bauru E um q fez em Guarulhos
# Eventos não excludentes, independentes.
probability_intersection_event(len(dataset_bauru), len(dataset_guarulhos), len(dataset_sp))

# Dados prob. de tirar x E y em seq.
probability_intersection_event(1, 1, 6)

# FN Probabilidade Condicional de Eventos A|B - SE (Dois Eventos Dependentes)
# a prob. é igual o produto dos eventos
# events_possible = número de Eventos possíveis dentro do espaço amostral
#  sample_space =  Espaço Amostral de todas possibilidades
def probability_conditional_event(events_possible, sample_space):
    result =  (events_possible / sample_space)*100
    print("{:.2f}%".format(result))
    return result

# Probabilidade e sortear uma Mulher Parda
dataset_woman = dataset_sp.loc[dataset_sp.SEXO == "F"]
dataset_woman_brown = dataset_woman.loc[dataset_woman.RACA == "parda"]
dataset_woman.shape, dataset_woman_brown.shape

# Probabilidade e sortear uma Mulher Parda
probability_conditional_event(len(dataset_woman_brown), len(dataset_woman))
