==> data_science_and_statistic_with_python/43-prob_distribuicao_binomial.ipynb <==
import pandas as pd
import numpy as np
from scipy.stats import binom

# POPULAÇÃO:
dataset_sp = pd.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
dataset_sp.shape

# Distribuição BINOMIAL: 
# Apenas dois resultados: Sucesso, Fracasso
# Tentativas idênticas e independentes
dataset_woman = dataset_sp.loc[dataset_sp.SEXO == "F"]
dataset_sp.shape

# Prob. de sortear uma mulher no enem SP
probability_woman = len(dataset_woman)/len(dataset_sp)
probability_woman

# binom.pmf (valor pontual)
# binom.cdf (faixa de valores acumulada do 0 até o ponto estipulado)

# Prob de sortear EXATAMENTE 4 mulheres em um total de 10 Amostras
# Primeiro parâmetro:valor ou limite que se está pretendo calcular
# Segundo: número de tentativas
# Terceiro: probabilidade de um sucesso.
result_1 = binom.pmf(4,10,probability_woman)
result_1 * 100

# Usando Regra de NÃO Ocorrência
# Prob de sortear PELO MENOS 1 mulheres em um total de 10 Amostras
# Se calcula na realidade a prob de NÃO sair nenhuma mulher (zero)
# 1 Menos pois é semelhante a prob da não ocorrência 
result_2 = 1 - binom.pmf(0,10,probability_woman)
result_2 * 100

# Usando Regra de NÃO Ocorrência
# Prob de sortear MAIS q 1 mulheres em um total de 10 Amostras
# Se calcula na realidade a prob de NÃO sair nenhuma mulher (zero)
# + a prob de não sair 1 única Mulher
# prob de não tirar + a prob de tirar apenas 1
result_3 = 1 - (binom.pmf(0,10,probability_woman)+binom.pmf(1,10,probability_woman))
result_3 * 100

# Usando Regra de NÃO Ocorrência
# Prob de sortear MAIS q 5 mulheres em um total de 10 Amostras
# + a prob de não sair 1 única Mulher
# prob de não tirar + a soma das demais prob.
result_4 = 1 - (binom.pmf(0,10,probability_woman)+binom.pmf(1,10,probability_woman)
                +binom.pmf(2,10,probability_woman)+binom.pmf(3,10,probability_woman)
                +binom.pmf(4,10,probability_woman)+binom.pmf(5,10,probability_woman))
result_4 * 100

# OU
# Usando a soma das Prob. de Ocorrência
# Prob de sortear MAIS q 5 mulheres em um total de 10 Amostras
# + a prob de não sair 1 única Mulher
result_5 = (binom.pmf(10,10,probability_woman)+binom.pmf(9,10,probability_woman)
            +binom.pmf(8,10,probability_woman)+binom.pmf(7,10,probability_woman)
            +binom.pmf(6,10,probability_woman))
result_5 * 100

# OU
# usando o "cdf" Acumula as prob. do 0 até o 5
# Usando Regra de NÃO Ocorrência
# Prob de sortear MAIS q 5 mulheres em um total de 10 Amostras
result_6 = 1 - binom.cdf(5,10,probability_woman)
result_6 * 100

# Usando Regra de Ocorrência
# Prob de sortear MAIS q 8 mulheres em um total de 10 Amostras
result_7 = binom.pmf(10,10,probability_woman)+binom.pmf(9,10,probability_woman)
result_7 * 100

# OU
# Prob de sortear MAIS q 8 mulheres em um total de 10 Amostras
# Calcula a prob de 10 acumulada e diminui a de 8 acum.
result_8 = binom.cdf(10,10,probability_woman) - binom.cdf(8,10,probability_woman)
result_8 * 100
