==> data_science_and_statistic_with_python/43-prob_distribuicao_discreta_dado_moeda.ipynb <==
import pandas as pd
import numpy as np
from scipy.stats import binom

# POPULAÇÃO:
dataset_sp = pd.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
dataset_sp.shape

# binom.pmf (valor pontual)
# binom.cdf (faixa de valores acumulada)

# Dados prob de tirar o número "x" em 1 lançamento
probaility_dice = 1/6
probaility_dice * 100

# Dados prob de tirar o número "x" EXATAMENTE 3 vezes em 10 lançamentos
result_dice_1 = binom.pmf(3,10, probaility_dice)
result_dice_1*100

# Dados prob de tirar o número "x" EXATAMENTE 5 vezes em 10 lançamentos
result_dice_2 = binom.pmf(5, 10, probaility_dice)
result_dice_2*100

# Usando Regra da Não Ocorrência
# Dados prob de tirar o número 1 ao menos 1 vez em 10 lançamentos
result_dice_3 = 1 - binom.pmf(0, 10, probaility_dice)
result_dice_3*100

# Moeda prob de tirar uma face em 1 lançamento
probaility_coin = 1/2
probaility_coin * 100

# Moeda prob de tirar face EXATAMENTE 1 vez em 2 lançamento
result_coin_4 = binom.pmf(1, 2, probaility_coin)
result_coin_4*100

# Moeda prob de tirar a mesma face EXATAMENTE 5 vezes em 10 lançamento
result_coin_3 = binom.pmf(5, 10, probaility_coin)
result_coin_3*100

# Moeda prob de tirar a mesma face EXATAMENTE 3 vezes em 10 lançamento
result_coin_1 = binom.pmf(3, 10, probaility_coin)
result_coin_1*100

# Moeda prob de tirar a mesma face EXATAMENTE 7 vezes em 10 lançamento
result_coin_2 = binom.pmf(7, 10, probaility_coin)
result_coin_2*100

# Moeda prob de tirar a mesma face EXATAMENTE 50 vezes em 100 lançamento
result_coin_5 = binom.pmf(50, 100, probaility_coin)
result_coin_5*100

# Moeda prob de tirar a mesma face EXATAMENTE 500 vezes em 1000 lançamento
result_coin_6 = binom.pmf(500, 1000, probaility_coin)
result_coin_6*100
