==> data_science_and_statistic_with_python/43-prob_distribuicao_geometrica.ipynb <==
import pandas as pd
import numpy as np
from scipy.stats import binom
from scipy.stats import geom
from scipy.stats import poisson

# POPULAÇÃO:
dataset_sp = pd.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
dataset_sp.shape

# Distribuição GEOMÉTRICA
# Prob. de repetidas tentativas até q o sucesso ocorra.
# Tentativas independentes
# A Prob de Sucesso é constante para todas tentativas
dataset_woman = dataset_sp.loc[dataset_sp.SEXO == "F"]
dataset_sp.shape

# Prob. de sortear uma mulher no enem SP
probability_woman = len(dataset_woman)/len(dataset_sp)
probability_woman

# x representa a tentativa de se obter sucesso
# p a probabilidade
# geom.pmf(x, p)

# Prob. de sortear 3 amostras e nenhuma ser mulher
# a 4ª tentativa seria o sucesso
result_1 = geom.pmf(4, probability_woman)
result_1 * 100
