==> data_science_and_statistic_with_python/43-prob_distribuicao_poisson.ipynb <==
import pandas as pd
import numpy as np
from scipy.stats import binom
from scipy.stats import geom
from scipy.stats import poisson

# Distribuição POISSON
# Prob. de vezes de um evento ocorrer unum determinado intervalo
# Prob. do evento ocorrer é a mesma para cada intervalo
# Núm. de ocorrências nos intervalos são independentes

# Num local de prova, 100 vestibulandos, normalmente, terminaram a prova em 2 horas (tempo mínimo).
# Probabilidade de exatamente 90 vestibulandos terminarem a prova em 2h.

# poisson.pmf(x, m) x é a quantidade de ocorrências EM ESTUDO e m é a taxa de ocorrências.

result_1 = poisson.pmf(90,100)
result_1 * 100
