==> data_science_and_statistic_with_python/49-teste_hipotese_teste_z.ipynb <==
import pandas as pd
import numpy as np
import plotly.express as px
import seaborn as sns
import scipy.stats as stats
import statsmodels
from statsmodels.stats.diagnostic import lilliefors
import matplotlib.pyplot as plt

# POPULAÇÃO:
dataset_sp = pd.read_csv("csv/04-classificados_tratado.csv", sep=",", encoding="iso-8859-1")
dataset_sp.shape
